feat!: Reorganize string functions
- Move string functions into `kolmafia-string.d.ts`
  (this does not affect users)
- Deprecate all string functions that can be replaced with JavaScript
  builtins
- Remove all functions that take an ASH buffer as argument, as they are
  effectively unusable from JavaScript.
  This includes:
  - `append()`
  - ~~`delete()`~~ was never exported in the first place
  - `insert()`
  - `replace()`
  - `setLength()`
- Add JSDoc comments for string functions
