+++ src/index.d.ts
@@ -3,3 +3,4 @@ export * from './kolmafia';
 export * from './kolmafia-cargo-cultist-shorts';
 export * from './kolmafia-math';
 export * from './kolmafia-modifier';
+export * from './kolmafia-string';
